requirements.txt
langchain-community
openai
mysql-connector-python==8.4.0
openpyxl
pandasai
pillow
streamlit
pandas